deleting an image in the folder too

+++ controllers/admin.js
@@ -1,5 +1,7 @@
+const fileHelper = require('../util/file');
 const {validationResult} = require('express-validator');
 const Product = require('../models/product');
+const product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', {
@@ -147,6 +149,7 @@ exports.getAddProduct = (req, res, next) => {
         product.price = updatedPrice;
         product.description = updatedDesc;
         if(image){
+          fileHelper.deleteFile(product.imageUrl);
           product.imageUrl = image.path;
         }
         
@@ -184,7 +187,14 @@ exports.getAddProduct = (req, res, next) => {
 
   exports.postDeleteProduct=(req,res,next)=>{
     const prodId = req.body.productId;
-    Product.deleteOne({_id:prodId, userId: req.user._id})
+    Product.findById(prodId)
+    .then(product=>{
+      if(!product){
+        return next(new Error('Product not found.'))
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({_id:prodId, userId: req.user._id})
+    })
     .then(()=>{
       console.log('destroyed product');
       res.redirect('/admin/products');
